refactor: mv DataSourceInfo out of mod.rs.
